# Improve the performance of StringBuilder.Append(char)