This PR is not complete yet, I need to check how Tiered JIT affects generated ASM with VTune.

By enforcing inlining of the `StringBuilder.Append(char)` I already got a nice improvement:

```cs
const int length = 50000;

[Benchmark]
public string StringBuilderAddChar()
{
    var result = new StringBuilder();

    for (int i = 0; i < length; i++)
        result.Append('s');

    return result.ToString();
}

[Benchmark]
public string StringBuilderAddCharOptimized()
{
    var result = new StringBuilder(length);

    for (int i = 0; i < length; i++)
        result.Append('s');

    return result.ToString();
}
```

|                        Method |                     Toolchain |     Mean | Ratio |
|------------------------------ |------------------------------ |---------:|------:|
|          StringBuilderAddChar |  \after\Core_Root\CoreRun.exe | 119.8 us |  0.92 |
|          StringBuilderAddChar | \before\Core_Root\CoreRun.exe | 130.8 us |  1.00 |
|                               |                               |          |       |
| StringBuilderAddCharOptimized |  \after\Core_Root\CoreRun.exe | 114.2 us |  0.83 |
| StringBuilderAddCharOptimized | \before\Core_Root\CoreRun.exe | 138.1 us |  1.00 |

Fixes #26825 #25774