@EgorBo 2019-10-03
@adamsitnik btw, I think you can also optimize the rangecheck inside that Append https://github.com/dotnet/coreclr/blob/1c5f18dde35784c2dd607e37803798c57973287b/src/System.Private.CoreLib/shared/System/Text/StringBuilder.cs#L1158 (currently it does emit a bound check)