@gfoidl 2019-10-03
@EgorBo 

```diff
- m_ChunkLength++;
+ m_ChunkLength = len + 1;
```
as it's already in the register.