@jkotas 2019-10-03
What is the root cause of the regression compared to .NET Framework? .NET Framework did not have this method with AggressiveInlining, so something regressed in .NET Core.