@adamsitnik 2019-10-15
@jkotas this PR is a draft, I am experimenting with benchmarks and profiler locally and using CI to run the tests for me. When it's ready I am going to ask you for a proper review. Currently, you are reviewing my draft ideas that most probably won't be part of the final version.