@jkotas 2019-10-15
The purpose of draft PRs is to [start conversation](https://github.blog/2019-02-14-introducing-draft-pull-requests/). If you explicitly do not want anybody to comment on the change, mark it as `[do not review]` or something similar.