@stephentoub 2019-10-15
And consider the cost involved in using CI to run all tests across all legs when StringBuilder can be 99.9% be validated locally just by running its tests locally.