@danmosemsft 2019-10-19
> What is the root cause of the regression compared to .NET Framework? .NET Framework did not have this method marked with AggressiveInlining, so something regressed in .NET Core.

Wonder what the perf numbers are if you simply paste in the.NEtFramework implementation of StringBuilder. Does the regression improve?